tests/wpt/web-platform-tests/css/css-animations/animation-iteration-count-calc.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Test: Animation count accepts calc()</title>
<link rel="author" name="Emilio Cobos Ãlvarez" href="mailto:emilio@crisal.io">
<link rel="help" href="https://drafts.csswg.org/css-values-3/#number-value">
<link rel="help" href="https://drafts.csswg.org/css-animations/#animation-iteration-count">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="target" style="animation-iteration-count: calc(1 + 3)">
<script>
test(function() {
  let target = document.getElementById("target");
  assert_equals(getComputedStyle(target).animationIterationCount, "4")
}, "calc() should be accepted in animation-iteration-count.");
</script>
